Setting Westmarcheswiki to allow external images.
We wish to display external images and other media on our site as to comply with copyright law for copyrighted material. Our admin, Ithiloneth ahs spoken to a +JohnLewis on the Miraheze IRC about this to get a feel for whether it might be appropriate or not before asking for this pull.