Disable BootstrapMediaWiki
So I was playing around with the theme, created the template Bootstrap:Subnav and then nothing works! It always give me a 500 error and I don't wont to crash the server. Even if I want to go back to the settings page and change my theme to the default one, I can't!