Remove commons.wikimedia.org cache (#2777)
This will improve performance and also use less disk space. This is because it won't be generating thumberneils (and can fetch from commons which is faster). It'll also use less disk space as it won't be saving to disk. And lastly it'll use less bandwith for the mediawiki servers as the client will be fetching directly.