Remove TLS cert and key from Database.php (#3055)
Since day 1 Miraheze has been relying on passwords (PSKs) to authenticate the mediawiki and wikiadmin users for database connections. As such, there is no point in setting the certificate and key paths here, since we don't use two-way authentication. Avoids log spam, extra traffic, more usage of the wildcard certificate (ideally only used on the cache proxies!) and avoids the impression of using one-way authentication where we don't.

Two-way authentication has its benefits and may be considered in the future, but is not relevant right now.