Update LocalSettings.php
In the case of "wikitech," per discussion with @RhinosF1 on Discord following a declined wiki creation request. Though we may not need it imminently, it seems prudent to reserve this subdomain for potential Miraheze use as a public or private wiki in the future. Likewise, for "gazetteer," per the several declined wiki requests which @Pix1234 and I have declined, this one seeks to replicate our Gazetteer of Wikis page on Meta without some sort of community discussion and clearly articulating how we could not improve upon the data available in/provided by the WikiDiscover special page. We may want to wildcard "gazetteer" to include "gazeteer" as a misspelling.