Update LocalSettings.php
Adding "sandbox" to restricted subdomains list because we've had a couple wiki request declines in the past month (https://meta.miraheze.org/wiki/Special:RequestWikiQueue/12598 and https://meta.miraheze.org/wiki/Special:RequestWikiQueue/13380) and I don't want to see it inadvertently get created. Also, adding "noc" as a restricted subdomain as I didn't see it in this list, and it's highly conceivable Miraheze may want to use it for technical coordination purposes, either as a wiki or some other website. Finally, though I don't want to create it now, I was talking with Reception123 on Discord about ways we can (a) increase Meta participation and (b) community volunteerism, and one of the ways in which I think this would be helpful would be through an Outreach wiki, so am requesting we restrict this subdomain for now.