Update LocalSettings.php to add 'automotive' category to CreateWiki and fix typo in 'eletronics' category (#3243)
* Update LocalSettings.php

One of the wiki requests in the queue is categorized as 'electronics'. We don't have any sufficient categories for this type of wiki. Additionally, we have a number of automotive- and automobile-related wikis that could possibly be categorized, if so desired by the local bureaucrats on those wikis, with this added category. Also, fixing a modest typo in 'eletronics' to 'electronics', which likely doesn't require further updates. Pinging @paladox or @NDKilla to see if they can approve and merge. If not, @Reception123 can likely do in the morning when he's awake and caffeinated.

* Update LocalSettings.php

Let's see if this stops @travis-ci from squawking.