add static.miraheze.org as a default allowed value of $wgTemplateStylesAllowedUrls
Now, I think I should mention when it comes to regular expressions, I don't have as much of an idea of what I am doing, as such I am unsure if this is correct. I believe the first 2 are, but the final (svg) one seems like it is forcing the URL to be something like https://static.miraheze.org/something.svg but like I said, am uncertain.