Update wgMaxCredits for testwiki
Per https://www.mediawiki.org/wiki/Manual:$wgMaxCredits, I'd like to set this function to a negative integer to list all contributors to a given page in the page's footer, for nostalgia and testing purposes. I'm a Consul on TestWiki, so should be fine unless a consensus of other Consuls objects. :P