Add prueba to the CreateWiki subdomain blacklist
With Spanish Public Test Wiki finally hitting Special:DeletedWikis this morning, `prueba` should be added to the CreateWiki subdomain blacklist similar to `wikitest`, to prevent accidental creation of an unauthorized Public Test Wiki fork that did not have either (a) the local consensus of the TestWiki community or (b) consensus of the global Miraheze community on community noticeboard. Incorporates standard regex for other subdomains to prevent numerical digits being included in the subdomain.