Modifies one reason; re-adds another reason
With regard to the duplicate wiki request reason, now that I think about it, this seems to be a duplication of "Content Policy (makes it difficult for other wikis)" and may be too specific of a reason, to be honest. This may be a case where we'd be better of declining per the "Content Policy (makes it difficult for other wikis)" reason and adding an additional comment advising, specifically, what the issue is.