Clarify short name decline reason; add two additional CP reasons
In the case of the former, it's mainly per concision but also because the requestor doesn't //have// to have been a `bureaucrat`, especially in the case of //public// wikis. Ideally, they are the original wiki requestor, but this, too, isn't specifically required.

In the case of the former, though less common decline reasons, it is nonetheless useful to have these decline reasons added to the canned responses should they need to be used and so the CreateWiki AI can learn from and adapt to actioned wiki requests.