Add Art & Architecture category to CreateWiki
We do get quite a few wikis that propose to create either a wiki about buildings and other real world infrastructure. Combining with "Art", as is common, also means this could potentially be used for the many artist repository wikis or even fan-made original character repositories/catalogues, too. It's also a common categorization