Better canned responses (#3761)
* Better canned responses

* Update LocalSettings.php

* Add a new approval reason

* Change the order of responses (and make a change to the "Use Public Test Wiki" reason)

* Update LocalSettings.php

* A new response and a few updates

* Minor copyediting, clarification to phrasing, deletion of one reason

Regarding the duplicate wiki reason, I'm open to adding that, but we have never given the original wiki requestor a carte blanche monopoly on the same topic. We don't allow complete or substantial carbon copies of existing wikis.

* Add another response

* Modifies one reason; re-adds another reason

With regard to the duplicate wiki request reason, now that I think about it, this seems to be a duplication of "Content Policy (makes it difficult for other wikis)" and may be too specific of a reason, to be honest. This may be a case where we'd be better of declining per the "Content Policy (makes it difficult for other wikis)" reason and adding an additional comment advising, specifically, what the issue is.

* Amend changes per request (https://github.com/R4356th/mw-config/pull/3)

* Amend changes per (https://github.com/R4356th/mw-config/pull/3)

* Fix spacing

* Clarify short name decline reason; add two additional CP reasons

In the case of the former, it's mainly per concision but also because the requestor doesn't //have// to have been a `bureaucrat`, especially in the case of //public// wikis. Ideally, they are the original wiki requestor, but this, too, isn't specifically required.

In the case of the former, though less common decline reasons, it is nonetheless useful to have these decline reasons added to the canned responses should they need to be used and so the CreateWiki AI can learn from and adapt to actioned wiki requests.

* Fix punctuation for consistency

* Update LocalSettings.php

* Refine this message per discussion with R4356th on Discord

* Update LocalSettings.php

Co-authored-by: Doug Mehus <39928941+dmehus@users.noreply.github.com>
Co-authored-by: Universal Omega <54654040+Universal-Omega@users.noreply.github.com>