Fix case
One more fix to this sitenotice. The other thing I was wondering if for `Special:ManageWiki/settings`, I'm wondering if we could make that into a link with `<a href="https:" + $wgServer + "/wiki/Special:ManageWiki/settings">Special:ManageWiki/settings</a>? I just never know if we need a `+` before and/or after the `$wgServer` variable?