Enable MP3 for my site using TimedMediaHandler (#3928)
* Enable MP3 for my site using TimedMediaHandler

As the title says. Also, this option has not been implemented into managewiki (mp4 has), so this is the only way for me to enable this setting. Maybe consider adding it to managewiki?

* Added another category to be highlighted on my wiki

I needed to make this change as well so I thought I'd might as well do it here. Hope I'm not doing anything wrong