Re-add user rights to steward group
Per discussion with @RhinosF1 on IRC that it wasn't necessary for Stewards. Also, the user rights still exist for the `sysadmin` and `trustandsafety` global groups, so I'm not sure to what end this would solve. `createwiki` from `wikicreator` I can see, given the sheer number of wiki creators we have, and the problems that would be caused if a wiki were created during the migration.