Move globalblock to local group
With the caveat that Owen _may_ want to keep it global per the reason(s) I expressed above. If made local, there may be times when we want to add a global block off Meta Wiki, but we have the permissions to edit our global group, so can do that then. For now, this seems fine