Set spamblacklist to true on all wikis (#4633)
* #9127

Set `default` to `true`, as this would be useless to us otherwise. Individual wikis would still be able to control who has access to the spam blacklist log by setting `spamblacklist` to a group other than `user`

* Simplify

Since the `default` setting will now be `true` with this PR, we now only need to specify custom wiki settings when a wiki has specified a setting other than the default (i.e., a wiki wants it to be `false`).