FileBackend: remove shardViaHashLevels (#4957)
levels being zero seems to be equivalent to nothing at all. Anything
other than levels being zero broke everything, and this is not needed
for us unless we have very massive wikis, which none of our wikis needs
to use container sharding, and also I believe it requires additional
configuration to support it.