Modify $wgAvailableRights and $wgRestrictionLevels (#5210)
I want to modify these restriction levels for the Scratchpad Wiki here
on Miraheze and show them in order. I am also adding templateeditor,
extendedconfirmed, moderator, and bureaucrat permissions to
$wgAvailableRights. I understand that the '', 'user', 'autoconfirmed',
and 'sysop' groups are already defined by default, but I want them to be
in order on the wiki I'm administrating. Thank you.