Enable the Parsoid extension on all wikis (#5243)
Eventually, this will be our main parser. In preparation, let's make it
available for everyone.

For now, this will have little impact apart from making the rest API
open to all wikis. (Although you shouldn't be rest.php for bots as it
isn't stable). In future, wikis will be able to start testing Parsoid
for reads which will be coming to Miraheze in about 7 months time.