T11040: Create $wmgMirahezeContactPageFooter
This variable is used for a hook that adds a link to the wiki's Special:Contact page to the footer.

While in ManageWiki it is set as a variable from the ContactPage extension, it is not. It's set like that so that it will only appear if ContactPage is enabled

As of writing, I'm not sure of the naming convention or where the hook should be, so I set it to wmgMirahezeContactPageFooter and added it to LocalSettings.php, respectively. I don't see any examples to go off by anyway.