Allow *.{miraheze,wikitide}.org and the custom domain for TemplateStyles
Currently, /w/img_auth.php URLs cannot be used for TemplateStyles, which is
problematic for private wikis (example: #general in Discord/#miraheze-offtopic
on IRC, at 2024-09-17 10:17:35 UTC).

We allow `[a-zA-Z0-9\-]\.(miraheze|wikitide)\.org/w/img_auth.php` for Miraheze
controlled domains, and `{$wi->server}/w/img_auth.php` for custom domains. In
the event that the wiki is not a custom domain, it will fallback to the
Miraheze domain and would effectively be a no-op.