change rate limit to 2 wikis every 24 hours
The current limit doesn't make too much sense as there's no need for anyone to be requesting more than a wiki per hour. In case someone needs a "companion" wiki it makes more sense to allow users to request 2 wikis but only every 24 hours (which would still be a lot).