T13612: Add settings for NewUserMessage (#5932)
Adds all three settings
([docs](https://www.mediawiki.org/wiki/Extension:NewUserMessage#Configuration))
of the extension to ManageWiki which resolves
https://issue-tracker.miraheze.org/T13612
(`$wgNewUserMessageOnAutoCreate` was requested, I added the other two
ones as well since they weren't in ManageWiki either).

`$wgNewUserMessageOnAutoCreate` will have to be set to true for
`nmfwikiwiki` via ManageWiki when this is merged since this was
specified in LocalSettings before this change.